refactor: #95 - 훅 이름 수정
